% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/series.R
\name{addOutliers<-.X13Series}
\alias{addOutliers<-.X13Series}
\title{Add Outliers: ao, ls, tc, so.}
\usage{
\method{addOutliers}{X13Series}(x, arima_model, update_save, correct_spec, values) <- value
}
\arguments{
\item{x}{Object.}

\item{value}{a character vector of valid ao parameters}
}
\description{
This function needs to be generalised to addParameterVals
}
