% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group.R
\name{addOutliers<-.X13SeriesGroup}
\alias{addOutliers<-.X13SeriesGroup}
\title{Add Outliers to all series in a group}
\usage{
\method{addOutliers}{X13SeriesGroup}(x, arima_model, update_save, correct_spec, values) <- value
}
\arguments{
\item{values}{}
}
\value{

}
\description{
Add Outliers to all series in a group
}
