% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spec.R
\name{addOutliers<-.X13SpecList}
\alias{addOutliers<-.X13SpecList}
\title{Add Outliers: ao, ls, tc, so.}
\usage{
\method{addOutliers}{X13SpecList}(
  x,
  arima_model = "(0 1 1)(0 1 1)",
  update_save = TRUE,
  correct_spec = TRUE,
  values
) <- value
}
\arguments{
\item{x}{Object.}

\item{value}{a character vector of valid ao parameters}
}
\description{
This function needs to be generalised to addParameterVals
}
