% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/series.R
\name{addParamVals<-.X13Series}
\alias{addParamVals<-.X13Series}
\title{Add Parameter Values to a SpecList parameter}
\usage{
\method{addParamVals}{X13Series}(x, spec, parameter, values) <- value
}
\arguments{
\item{values}{}
}
\value{

}
\description{
Add Parameter Values to a SpecList parameter
}
