% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group.R
\name{addParamVals<-.X13SeriesGroup}
\alias{addParamVals<-.X13SeriesGroup}
\title{Add parameter values to all series in a group}
\usage{
\method{addParamVals}{X13SeriesGroup}(x, spec, parameter, values) <- value
}
\arguments{
\item{values}{}
}
\value{

}
\description{
Add parameter values to all series in a group
}
