% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/series.R
\name{adjust.X13Series}
\alias{adjust.X13Series}
\title{Perform seasonal adjustment.}
\usage{
\method{adjust}{X13Series}(x, purge = TRUE, keep_fixed = FALSE, grp_num = 1L, ...)
}
\arguments{
\item{x}{Object on which to perform adjustment.}
}
\description{
Perform seasonal adjustment with \code{X13-ARIMA-SEATS}.
}
