% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{as.data.frame.X13Diagnostics}
\alias{as.data.frame.X13Diagnostics}
\title{Coerce \code{X13Diagnostics} object to \code{\link{data.frame}}.}
\usage{
\method{as.data.frame}{X13Diagnostics}(
  x,
  statfilter,
  colfilter = c("stat", "value"),
  stringsAsFactors = FALSE
)
}
\description{
Coerce \code{X13Diagnostics} object to \code{\link{data.frame}}.
}
