% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spec.R
\name{correctARIMA.X13SpecList}
\alias{correctARIMA.X13SpecList}
\title{Correct a SpecList.}
\usage{
\method{correctARIMA}{X13SpecList}(x)
}
\arguments{
\item{x}{The Object to correct.}
}
\value{

}
\description{
For example, if the are regression variables in a multiplicative model,
 then add to transform funciton=log
}
