% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/series.R
\name{correctSeriesSpec.X13Series}
\alias{correctSeriesSpec.X13Series}
\title{Validates/corrects a series Spec file}
\usage{
\method{correctSeriesSpec}{X13Series}(x)
}
\arguments{
\item{x}{}
}
\value{

}
\description{
Validates/corrects a series Spec file
}
