% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{gather}
\alias{gather}
\title{Gather a list of values into a string, with linebreaks where required.}
\usage{
gather(xs, max = 132, indent = 4)
}
\description{
Gather a list of values into a string, with linebreaks where required.
}
\keyword{internal}
