% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch.R
\name{getParamVals.X13Batch}
\alias{getParamVals.X13Batch}
\title{Get Spec Parameter Values of Batch}
\usage{
\method{getParamVals}{X13Batch}(
  x,
  spec,
  parameter,
  snames = NULL,
  simplify = if_else(rlang::is_empty(snames), FALSE, TRUE)
)
}
\arguments{
\item{x}{}
}
\value{

}
\description{
Get Spec Parameter Values of Batch
}
