% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/series.R
\name{getParamVals.X13Series}
\alias{getParamVals.X13Series}
\title{Get Parameter Values}
\usage{
\method{getParamVals}{X13Series}(x, spec, parameter)
}
\arguments{
\item{x}{Object.}

\item{spec}{The spec name.}

\item{parameter}{The parameter name.}
}
\description{
Returns parameter values as a character vector
}
