% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch.R
\name{getPeriod.X13Batch}
\alias{getPeriod.X13Batch}
\title{Get period of series in the batch}
\usage{
\method{getPeriod}{X13Batch}(x)
}
\arguments{
\item{x}{}
}
\value{

}
\description{
Get period of series in the batch
}
