% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/series.R
\name{getSpec.X13Series}
\alias{getSpec.X13Series}
\title{Get a spec.}
\usage{
\method{getSpec}{X13Series}(x, spec)
}
\arguments{
\item{x}{Input object.}

\item{name}{Parameter name.}

\item{value}{Parameter value.}
}
\description{
Get a spec.
}
