% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spec.R
\name{getSpec.X13SpecList}
\alias{getSpec.X13SpecList}
\title{Get a spec.}
\usage{
\method{getSpec}{X13SpecList}(x, spec)
}
\arguments{
\item{x}{Input object.}

\item{spec}{Parameter name.}
}
\description{
Get a spec.
}
