% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/series.R
\name{getSpecList.X13Series}
\alias{getSpecList.X13Series}
\title{Get a specification list.}
\usage{
\method{getSpecList}{X13Series}(x)
}
\arguments{
\item{x}{Input object.}
}
\description{
Get a specification list.
}
