% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch.R
\name{getSpecParameter.X13Batch}
\alias{getSpecParameter.X13Batch}
\title{Get Spec Parameter of Batch}
\usage{
\method{getSpecParameter}{X13Batch}(
  x,
  spec,
  parameter,
  snames = NULL,
  simplify = if_else(rlang::is_empty(snames), FALSE, TRUE)
)
}
\arguments{
\item{x}{}
}
\value{

}
\description{
Get Spec Parameter of Batch
}
