% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spec.R
\name{getSpecParameter.X13SpecList}
\alias{getSpecParameter.X13SpecList}
\title{Get a spec parameter.}
\usage{
\method{getSpecParameter}{X13SpecList}(x, spec, parameter)
}
\arguments{
\item{x}{Input object.}

\item{spec}{Parameter name.}

\item{parameter}{Parameter value.}
}
\description{
Get the value of a spec parameter.
}
