% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spec.R
\name{\%+\%.X13Spec}
\alias{\%+\%.X13Spec}
\title{Combine \code{\link{X13Spec}} objects.}
\usage{
`\%+\%.X13Spec`(e1, e2)
}
\description{
Combine \code{\link{X13Spec}} objects.
}
\examples{
s1 <- X13Spec(specname = "x11", mode = "mult")

s2 <- X13Spec(
  specname = "x11", sigmalim = "(1.8,2.8)",
  save = "(d8 d10 d11 d12 d13 c17)"
)

s1 \%+\% s2
}
