% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spec.R
\name{\%+\%.X13SpecList}
\alias{\%+\%.X13SpecList}
\title{Combine \code{\link{X13SpecList}} objects, or add an \code{\link{X13Spec}}
object to an existing X13SpecList object.}
\usage{
\method{\%+\%}{X13SpecList}(e1, e2)
}
\description{
Combine \code{\link{X13SpecList}} objects, or add an \code{\link{X13Spec}}
object to an existing X13SpecList object.
}
\examples{
spec1 <- X13SpecList(series = list(start = "1949.1",
period = "12",
title = "AirPassengers",
file = "AirPassengers.dat",
format = "datevalue",
save = "(b1)"))

spec2 <- X13SpecList(x11 = list(mode = "mult",
                                sigmalim = "(1.8,2.8)",
                                save = "(d8 d10 d11 d12 d13 c17)"))

spec1 \%+\% spec2
spec1 \%+\% spec2 \%+\% X13Spec(specname = "transform", `function`="log")
}
