% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{HLFS}
\alias{HLFS}
\title{Labour force status by sex from Statistics New Zealand's Household Labour Force Survey.}
\format{
A data frame with columns:
\describe{
  \item{name}{Series name.}
  \item{sname}{Abbreviated series name.}
  \item{year}{Calendar year.}
  \item{period}{Quarter number, 1 through 4.}
  \item{date}{Date representation of quarter--last calendar day of each quarter.}
  \item{value}{Data value.}
}
}
\usage{
HLFS
}
\description{
Labour force status by sex from Statistics New Zealand's Household Labour Force Survey.
}
\keyword{datasets}
