% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mta.R
\name{mtaToX13Series}
\alias{mtaToX13Series}
\title{Convert mta group to a list of X13Series}
\usage{
mtaToX13Series(mta_grp, windows_fp = FALSE)
}
\description{
Convert mta group to a list of X13Series
}
\keyword{internal}
