% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.X13SeriesResult}
\alias{plot.X13SeriesResult}
\title{Plot adjustment results}
\usage{
\method{plot}{X13SeriesResult}(x, type = "B1D11D12", interactive = FALSE, ...)
}
\arguments{
\item{x}{\code{X13SeriesResult}}

\item{type}{one of \code{B1D11D12}, \code{decomp}, \code{D10}
, \code{D10D8}, \code{D11oD11}, \code{D12oD12} or \code{b1d11isa}}
}
\description{
Plot adjustment results
}
