% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{readSPC0}
\alias{readSPC0}
\title{Read SPC file.}
\usage{
readSPC0(fname)
}
\arguments{
\item{fname}{file name}
}
\description{
This is the original version, containing bugs.
It will be removed after readSPC has been tested fully.
}
\keyword{internal}
