% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/series.R
\name{removeFacFile.X13Series}
\alias{removeFacFile.X13Series}
\title{Remove a factor file.}
\usage{
\method{removeFacFile}{X13Series}(x)
}
\arguments{
\item{x}{Input object.}
}
\description{
Remove a factor file.
}
