% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch.R
\name{removeOutliers<-.X13Batch}
\alias{removeOutliers<-.X13Batch}
\title{Remove outliers from all groups in a batch}
\usage{
\method{removeOutliers}{X13Batch}(x, values) <- value
}
\arguments{
\item{values}{}
}
\value{

}
\description{
ao, ls, so, tc
}
