% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/series.R
\name{removeOutliers<-.X13Series}
\alias{removeOutliers<-.X13Series}
\title{Remove outlier variables: ao, ls, so, tc}
\usage{
\method{removeOutliers}{X13Series}(x, values) <- value
}
\arguments{
\item{values}{}
}
\value{

}
\description{
Remove outlier variables: ao, ls, so, tc
}
