% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group.R
\name{removeOutliers<-.X13SeriesGroup}
\alias{removeOutliers<-.X13SeriesGroup}
\title{Remove outliers from all series in a group.}
\usage{
\method{removeOutliers}{X13SeriesGroup}(x, values) <- value
}
\arguments{
\item{values}{}
}
\value{

}
\description{
ao, ls, so, tc
}
