% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outliers.R
\name{removeOutOfSpanReg}
\alias{removeOutOfSpanReg}
\title{Removes regression Variables out of Span}
\usage{
removeOutOfSpanReg(reg_vars, s_period, s_end_date)
}
\arguments{
\item{speriod}{}
}
\value{

}
\description{
Removes regression Variables out of Span
}
