% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/series.R
\name{removeParamVals<-.X13Series}
\alias{removeParamVals<-.X13Series}
\title{Remove parameter values}
\usage{
\method{removeParamVals}{X13Series}(x, spec, parameter, values) <- value
}
\arguments{
\item{values}{}
}
\value{

}
\description{
Remove parameter values
}
