% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group.R
\name{removeParamVals<-.X13SeriesGroup}
\alias{removeParamVals<-.X13SeriesGroup}
\title{Remove parameter values from all series in a group.}
\usage{
\method{removeParamVals}{X13SeriesGroup}(x, spec, parameter, values) <- value
}
\arguments{
\item{values}{}
}
\value{

}
\description{
Remove parameter values from all series in a group.
}
