% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spec.R
\name{removeParamVals<-.X13SpecList}
\alias{removeParamVals<-.X13SpecList}
\title{Remove Parameter Values}
\usage{
\method{removeParamVals}{X13SpecList}(x, spec, parameter, values) <- value
}
\arguments{
\item{values}{}
}
\value{

}
\description{
Remove Parameter Values
}
