% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rmdups}
\alias{rmdups}
\title{Replace blocks adjacent matching characters with single character in string.}
\usage{
rmdups(s, c = " ")
}
\description{
Replace blocks adjacent matching characters with single character in string.
}
\keyword{internal}
