% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch.R
\name{selectSeries.X13Batch}
\alias{selectSeries.X13Batch}
\title{Select unique series}
\usage{
\method{selectSeries}{X13Batch}(x, snames, simplify = if_else(missing(snames), FALSE, TRUE))
}
\arguments{
\item{snames}{}
}
\value{

}
\description{
Select unique series
}
