% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch.R
\name{selectSeries.X13BatchResult}
\alias{selectSeries.X13BatchResult}
\title{Select unique series}
\usage{
\method{selectSeries}{X13BatchResult}(x, snames)
}
\arguments{
\item{snames}{}
}
\value{

}
\description{
Select unique series
}
