% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/series.R
\name{setFacFile<-.X13Series}
\alias{setFacFile<-.X13Series}
\title{Set a factor file.}
\usage{
\method{setFacFile}{X13Series}(x) <- value
}
\arguments{
\item{x}{Input object.}
}
\description{
Set a factor file.
}
