% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/series.R
\name{setParamVals<-.X13Series}
\alias{setParamVals<-.X13Series}
\title{Set Parameter Values.}
\usage{
\method{setParamVals}{X13Series}(x, spec, parameter, values) <- value
}
\arguments{
\item{x}{Object to modify.}

\item{spec}{The spec name.}

\item{parameter}{The parameter name.}

\item{values}{A character vector of Regression variables.}
}
\description{
Allows the user to set multiple parameter values with a character vector.
}
