% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/series.R
\name{setRegFile<-.X13Series}
\alias{setRegFile<-.X13Series}
\title{Set a regression file.}
\usage{
\method{setRegFile}{X13Series}(x) <- value
}
\arguments{
\item{x}{Input object.}
}
\description{
Set a regression file.
}
