% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch.R
\name{setSpecParameter<-.X13Batch}
\alias{setSpecParameter<-.X13Batch}
\title{Set parameter value for all series in a group.}
\usage{
\method{setSpecParameter}{X13Batch}(x, spec, parameter, snames = NULL) <- value
}
\arguments{
\item{values}{}
}
\value{

}
\description{
Set parameter value for all series in a group.
}
