% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group.R
\name{setSpecParameter<-.X13SeriesGroup}
\alias{setSpecParameter<-.X13SeriesGroup}
\title{Set parameter value for all series in a group.}
\usage{
\method{setSpecParameter}{X13SeriesGroup}(x, spec, parameter, snames = NULL) <- value
}
\arguments{
\item{values}{}
}
\value{

}
\description{
Set parameter value for all series in a group.
}
