% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spec.R
\name{setSpecParameter<-.X13Spec}
\alias{setSpecParameter<-.X13Spec}
\title{Set a spec parameter.}
\usage{
\method{setSpecParameter}{X13Spec}(x, name) <- value
}
\arguments{
\item{x}{Object to modify.}

\item{name}{Parameter name.}

\item{value}{Parameter value.}
}
\value{
An updated version of input \code{x}.
}
\description{
Set a spec parameter.
}
\examples{
spec <- X13Spec(
  specname = "x11", mode = "mult", sigmalim = "(1.8,2.8)",
  save = "(d8 d10 d11 d12 d13 c17)"
)

spec
setSpecParameter(spec, "mode") <- "add"
spec
setSpecParameter(spec, "mode") <- NULL
spec
}
