% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spec.R
\name{setSpecParameter<-.X13SpecList}
\alias{setSpecParameter<-.X13SpecList}
\title{Set a spec parameter.}
\usage{
\method{setSpecParameter}{X13SpecList}(x, spec, parameter) <- value
}
\arguments{
\item{x}{Object to modify.}

\item{parameter}{Parameter name}

\item{value}{Parameter value.}

\item{name}{Spec name.}
}
\value{
An updated version of input \code{x}.
}
\description{
Set a spec parameter.
}
