% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/series.R
\name{spanDate.X13Series}
\alias{spanDate.X13Series}
\title{Get the start or end date of the span parameter}
\usage{
\method{spanDate}{X13Series}(x, value = c("start"), spec = "series")
}
\arguments{
\item{value}{Can take "start" or "end"}
}
\description{
Get the start or end date of the span parameter
}
