% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/series.R
\name{spanStartDate.X13Series}
\alias{spanStartDate.X13Series}
\title{Get the start date of the span parameter}
\usage{
\method{spanStartDate}{X13Series}(x, spec = "series")
}
\description{
Get the start date of the span parameter
}
