% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpecificationParser.R
\docType{data}
\name{SPCparser}
\alias{SPCparser}
\title{emulate a singleton object}
\format{
An object of class \code{list} of length 8.
}
\usage{
SPCparser
}
\description{
emulate a singleton object
}
\keyword{internal}
