% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch.R
\name{summary.X13Batch}
\alias{summary.X13Batch}
\title{Summarise an X13 Batch object}
\usage{
\method{summary}{X13Batch}(x, ...)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Summarise an X13 Batch object
}
