% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch.R
\name{summary.X13BatchResult}
\alias{summary.X13BatchResult}
\title{Summarise an X13 BatchResult object}
\usage{
\method{summary}{X13BatchResult}(x, ...)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Summarise an X13 BatchResult object
}
