% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{tsdf}
\alias{tsdf}
\title{Convert time series object to a suitable data frame.}
\usage{
tsdf(x)
}
\arguments{
\item{x}{Object of class \code{\link{ts}}.}
}
\description{
Convert time series object to a suitable data frame.
}
\examples{
ap <- tsdf(AirPassengers)
ap
}
