% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch.R
\name{tvals.X13BatchResult}
\alias{tvals.X13BatchResult}
\title{Extract t-vals from regression variables}
\usage{
\method{tvals}{X13BatchResult}(x, otl_types = c("ao", "tc", "ls", "rp"), include_auto = TRUE)
}
\arguments{
\item{variables}{}
}
\value{

}
\description{
Extract t-vals from regression variables
}
