% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch.R
\name{udg.X13BatchResult}
\alias{udg.X13BatchResult}
\title{Return UDG's of series in a batch.}
\usage{
\method{udg}{X13BatchResult}(x, snames = NULL, simplify = if_else(rlang::is_empty(snames), FALSE, TRUE))
}
\arguments{
\item{x}{}
}
\value{

}
\description{
Return UDG's of series in a batch.
}
