% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/series.R
\name{udg.X13SeriesResult}
\alias{udg.X13SeriesResult}
\title{Get the UDG (summary diagnostics)}
\usage{
\method{udg}{X13SeriesResult}(x)
}
\arguments{
\item{x}{}
}
\value{
A list representing the values in the UDG file
}
\description{
Get the UDG (summary diagnostics)
}
