% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StringUtils.R, R/utils.R
\name{unquote}
\alias{unquote}
\title{unquote}
\usage{
unquote(s)

unquote(s)
}
\arguments{
\item{s:}{String}
}
\value{
String
}
\description{
unquote

Remove quotes from string.
}
\keyword{internal}
