% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view.R
\name{view.X13SeriesGroupResult}
\alias{view.X13SeriesGroupResult}
\title{Show result of adjustment as interactive application.}
\usage{
\method{view}{X13SeriesGroupResult}(x)
}
\arguments{
\item{x}{{{X13SeriesGroupResult}} resulting from adjusting {{X13SeriesGroup}}}
}
\description{
Show result of adjustment as interactive application.
}
