% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view.R
\name{view.X13SeriesResult}
\alias{view.X13SeriesResult}
\title{Show result of adjustment as interactive application.}
\usage{
\method{view}{X13SeriesResult}(x)
}
\arguments{
\item{x}{{{X13SeriesResult}} resulting from adjusting {{X13Series}}}
}
\description{
Show result of adjustment as interactive application.
}
