% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spec.R
\name{writeSpecToFile.X13SpecList}
\alias{writeSpecToFile.X13SpecList}
\title{Write a spec to file}
\usage{
\method{writeSpecToFile}{X13SpecList}(x, fname)
}
\arguments{
\item{x}{The spec to write out.}

\item{fname}{The full file name.}
}
\description{
Write a spec to file
}
