% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/series.R
\name{X11AddMult.X13Series}
\alias{X11AddMult.X13Series}
\title{Is the x-11 method additive or multiplicative?}
\usage{
\method{X11AddMult}{X13Series}(x)
}
\arguments{
\item{x}{}
}
\value{
"add", "mult", or NA
}
\description{
Is the x-11 method additive or multiplicative?
}
