% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch.R
\name{X13BatchFromMTA}
\alias{X13BatchFromMTA}
\title{Create a list of X13SeriesGroup objects from an mta file.}
\usage{
X13BatchFromMTA(mta_path)
}
\arguments{
\item{mta}{Path to mta file.}
}
\description{
Create a list of X13SeriesGroup objects from an mta file.
}
