% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mta.R
\name{X13ListToGroup}
\alias{X13ListToGroup}
\title{Convert mta group to a list of X13Series}
\usage{
X13ListToGroup(ser_list, sname = "1")
}
\description{
Convert mta group to a list of X13Series
}
\keyword{internal}
