% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/series.R
\name{X13Series}
\alias{X13Series}
\title{Create an object to be adjusted by X13-ARIMA-SEATS.}
\usage{
X13Series(
  x,
  value = "value",
  sname = deparse(substitute(x)),
  lname = sname,
  type = "x11",
  speclist,
  facfile,
  ...
)
}
\arguments{
\item{x}{A data frame containing a single time series--should have columns
\code{year}, \code{period}, \code{value}.  If \code{x} is of class
\code{\link{ts}}, then it will be converted to the required format.}

\item{sname}{A short name for the series.}

\item{lname}{A long name for the series.}

\item{type}{Either of \code{x11} or \code{seats}.  When \code{speclist} is
left missing, a default is provided; and this determines whether the
provided specification contains an \code{x11} or \code{seats} block.}

\item{speclist}{A spec list created with \code{\link{X13SpecList}}.}

\item{facfile}{A factor file created with \code{\link{new_facFile}}.}

\item{...}{As an alternative to \code{SpecList}, a variable number of
name-value pairs.}
}
\description{
Create an object to be adjusted by X13-ARIMA-SEATS.
}
