% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group.R
\name{X13SeriesGroup}
\alias{X13SeriesGroup}
\title{Create an object to be adjusted by X13-ARIMA-SEATS.}
\usage{
X13SeriesGroup(sname, ..., lname = sname)
}
\arguments{
\item{sname}{A short name for the series.}

\item{...}{a list of \code{X13Series} objects}

\item{lname}{A long name for the series.}
}
\description{
Create an object to be adjusted by X13-ARIMA-SEATS.
}
