% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spec.R
\name{X13Spec}
\alias{X13Spec}
\title{X13-ARIMA-SEATS specification.}
\usage{
X13Spec(specname = "series", ...)
}
\arguments{
\item{...}{Variable number of name-value pairs, e.g. \code{mode = "mult"},
\code{mode = "sigmalim"}, etc.}

\item{name}{of specification, e.g. \code{series}, \code{x11}, \code{outlier},
\code{composite}, etc.}
}
\value{
A list of class \code{\link{X13Spec}}.  The first element,
\code{name}, is the name of the specification.  The second element,
\code{args}, is a list of name-value pairs.
}
\description{
List containing a single specification group, e.g. \code{series},
\code{x11}, \code{outlier}, etc.
}
\examples{
spec <- X13Spec(specname = "x11",
                mode = "mult",
                sigmalim = "(1.8,2.8)",
                save = "(d8 d10 d11 d12 d13 c17)")
spec
}
