% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{as.data.frame.X13Diagnostics}
\alias{as.data.frame.X13Diagnostics}
\title{Coerce \code{X13Diagnostics} object to \code{\link{data.frame}}.}
\usage{
\method{as.data.frame}{X13Diagnostics}(
  x,
  statfilter = c("f2.mcd", "f2.ic", "f2.is", "f2.fsd8", "f2.msf", sprintf("f3.m\%02d",
    1:11), "f3.q", "f3.qm2"),
  colfilter = c("stat", "value")
)
}
\description{
Coerce \code{X13Diagnostics} object to \code{\link{data.frame}}.
}
