% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spec.R
\name{getSpecParameter.X13Spec}
\alias{getSpecParameter.X13Spec}
\title{Get a spec parameter.}
\usage{
\method{getSpecParameter}{X13Spec}(x, spec, parameter)
}
\arguments{
\item{x}{Input object.}

\item{name}{Parameter name.}

\item{value}{Parameter value.}
}
\description{
Get the value of a spec parameter.
}
