% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpecificationParser.R
\name{SpecificationParser}
\alias{SpecificationParser}
\title{A closure of Functions to parse an X13 .spc file
parseSPC is the main function and its output
gets passed to parsedSpectoX13Spec which creates X13 classes
or alternatively converted to some other type like a JSON file}
\usage{
SpecificationParser()
}
\arguments{
\item{rs}{list of regular expression strings}
}
\value{
A regular expression string
}
\description{
A closure of Functions to parse an X13 .spc file
parseSPC is the main function and its output
gets passed to parsedSpectoX13Spec which creates X13 classes
or alternatively converted to some other type like a JSON file
}
\keyword{'any}
\keyword{Make}
\keyword{a}
\keyword{characters}
\keyword{concatenate}
\keyword{drop}
\keyword{first}
\keyword{internal}
\keyword{matches}
\keyword{of}
\keyword{of'}
\keyword{others.}
\keyword{regex}
\keyword{several}
\keyword{single}
\keyword{vector}
\keyword{which}
\keyword{x}
