% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpecificationParser.R
\name{SpecificationParser}
\alias{SpecificationParser}
\title{A closure of functions used to parse an X13 .spc file.}
\usage{
SpecificationParser()
}
\value{
A list of functions useful for parsing spec files.
}
\description{
\code{parseSPC} pipes the child functions together. Its output
  gets passed to parsedSpectoX13Spec which creates X13 classes.
}
\section{stripComment}{


Remove comments from a specification. Takes mapped output from readLines
Comments are denoted by #.  Remove any lines that start with #, or remove
content after # if it appears part-way through a line.
\subsection{Arguments}{
\describe{
  \item{str}{The body of a SPC file.}
}}
\subsection{Value}{A 1-element character: a SPC file line with comments removed.}
}

\section{preprocess}{

Remove newlines, remove comments, and squish whitespace from the content
  of a SPC file.
\subsection{Arguments}{
\describe{
  \item{str}{multiple-length chatacter: The body of a SPC file.}
}}
\subsection{Value}{A character vector of length 1 containing a tidier
  , squished version of the body of a SPC file.}
}

\section{parseSpecs}{

Split body of SPC files into individual specs:

Extract instances of the pattern spec \{ stuff \}, where spec is
  , for example, \code{series}, \code{x11}, etc.

\subsection{Arguments}{
\describe{
  \item{str}{The output of preprocess: a character vector of length 1.}
}}
\subsection{Value}{A character vector with length equal to the number of specs.}
}

\section{parseSpecNameAndBody}{

Convert a list of specifications a named list of values.

\subsection{Arguments}{
\describe{
  \item{str}{The output of parseSpecs:
    A character vector with length equal to the number of specs.}
}}
\subsection{Value}{A named list of values.}
}

\section{parseArgs}{

Parse Arguments:
  The output of parseSpecNameAndBody maps to this function.
\subsection{Arguments}{
\describe{
  \item{str}{A 1-element character containing raw spec values.}
}}
\subsection{Value}{A named list of named parameters.}
}

\keyword{internal}
